+++ bluetooth_low_energy_platform_interface/lib/src/central_manager.dart
@@ -1,6 +1,7 @@
 import 'dart:typed_data';
 
-import 'bluetooth_low_energy.dart';
+import 'package:plugin_platform_interface/plugin_platform_interface.dart';
+
 import 'bluetooth_low_energy_manager.dart';
 import 'event_args.dart';
 import 'gatt_characteristic.dart';
@@ -10,10 +11,33 @@ import 'gatt_service.dart';
 import 'peripheral.dart';
 
 /// An object that scans for, discovers, connects to, and manages peripherals.
-abstract class CentralManager extends BluetoothLowEnergyManager {
-  /// Gets the instance of [CentralManager].
-  static CentralManager get instance =>
-      BluetoothLowEnergy.instance.centralManager;
+abstract class CentralManager extends PlatformInterface
+    implements BluetoothLowEnergyManager {
+  /// Constructs a [CentralManager].
+  CentralManager() : super(token: _token);
+
+  static final Object _token = Object();
+
+  static CentralManager? _instance;
+
+  /// The default instance of [CentralManager] to use.
+  static CentralManager get instance {
+    final instance = _instance;
+    if (instance == null) {
+      throw UnimplementedError(
+        'CentralManager is not implemented on this platform.',
+      );
+    }
+    return instance;
+  }
+
+  /// Platform-specific implementations should set this with their own
+  /// platform-specific class that extends [CentralManager] when
+  /// they register themselves.
+  static set instance(CentralManager instance) {
+    PlatformInterface.verifyToken(instance, _token);
+    _instance = instance;
+  }
 
   /// Tells the central manager discovered a peripheral while scanning for devices.
   Stream<DiscoveredEventArgs> get discovered;