+++ bluetooth_low_energy_platform_interface/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 4.0.0-dev.1
+
+* Remove `BluetoothLowEnergy` class.
+* Update `CentralManger` to extends `PlatformInterface`.
+* Update `PeripheralManager` to extends `PlatformInterface`.
+* Update `README.md`.
+
 ## 3.0.0
 
 * Add `PeripheralManager` api.